requirements.txt
python-telegram-bot==20.2
requests==2.27.1